+++ cpp/liquid/src/ml_lcm.c
@@ -1,6 +1,7 @@
 #include "math-lib.h"
 
-double ml_lcm(double n1,double n2)
+int ml_lcm(int n1, int n2)
+// Compute the LCM(n1, n2)
 {
     return 0;
 }
