+++ cpp/liquid/src/ml_mod.c
@@ -1,6 +1,8 @@
 #include "math-lib.h"
 
-double ml_mod(double n1, double n2)
+int ml_mod(int n1, int n2)
+// Compute the remainder when n1 divides n2 (n1 % n2)
+// You cannot use the % operator.
 {
     return 0;
 }
